Develop

+++ src/public/user/_controllers/user.controller.ts
@@ -9,6 +9,7 @@ import {
   NotFoundException,
   UseGuards,
   Put,
+  ForbiddenException,
 } from '@nestjs/common';
 import { AuthGuard } from 'src/public/auth/_guards/auth.guard';
 import { UserService } from '../_services/user.service';
@@ -18,6 +19,7 @@ import { RuleAccess } from 'src/public/_decorators/rule-access.decorator';
 import { RuleAccessEnum } from 'src/public/_enums/rule-access.enum';
 import { RuleAccessGuard } from 'src/public/auth/_guards/rule-access.guard';
 import { ValidateAndTransformUsers } from '../_decorators/validate-and-transform-users.decorator';
+import { userDecorator } from 'src/public/_decorators/user.decorator';
 
 @Controller('user')
 export class UserController {
@@ -101,8 +103,15 @@ export class UserController {
   @HttpCode(HttpStatus.OK)
   async updateUser(
     @ParamNumberId() userId: number,
+    @userDecorator('id') currentUserId: number,
     @Body() user: Prisma.UserUpdateInput,
   ): Promise<{ statusCode: number; message: string }> {
+    if (userId !== currentUserId) {
+      throw new ForbiddenException(
+        'You are not authorized to update this user.',
+      );
+    }
+
     try {
       await this.userService.updateUsers(userId, user);
 
@@ -121,7 +130,14 @@ export class UserController {
   @HttpCode(HttpStatus.OK)
   async deleteUser(
     @ParamNumberId() id: number,
+    @userDecorator('id') currentUserId: number,
   ): Promise<{ statusCode: number; message: string }> {
+    if (id !== currentUserId) {
+      throw new ForbiddenException(
+        'You are not authorized to delete this user.',
+      );
+    }
+
     try {
       const deletedUser = await this.userService.deleteUser(id);
